### 2020-11-23 #95 @whedon
[PRE REVIEW]: StarBLAST: a scalable BLAST+ solution for the classroom
**Submitting author:** @cosimichele (<a href="http://orcid.org/0000-0001-7609-1939">Michele Cosi</a>)
**Repository:** <a href="https://github.com/LyonsLab/StarBLAST/" target ="_blank">https://github.com/LyonsLab/StarBLAST/</a>
**Version:** v1.0.0

@cosimichele if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/jose-reviewers) have already agreed to review for JOSE and may be suitable for this submission.
