### 2020-11-24 #95 @labarba
hi @CosiMichele — we expect a section titled "Statement of Need" in the paper. See our [documentation](https://openjournals.readthedocs.io/en/jose/submitting.html#what-should-my-paper-contain).

@AstrobioMike has agreed to be the first reviewer. We'll be looking for a second reviewer, assign a handling editor, and then start the REVIEW issue, where all the action happens. Stay tuned!
