### 2020-12-23 #95 @JasonJWilliamsNY
Hi  @CosiMichele I will be following up on this. Since we are at the beginning of the holidays reviewer invites will go out the first week in January.
