### 2021-01-13 #95 @JasonJWilliamsNY
@AstrobioMike back in November you volunteered as reviewer. I have just joined JOSE as and editor and @oneilish has agreed to review as well. Can you confirm you will be able to go ahead? Hope you two will meet as well!
