### 2021-01-14 #102 @whedon
[REVIEW]: StarBLAST: a scalable BLAST+ solution for the classroom
**Submitting author:** @cosimichele (<a href="http://orcid.org/0000-0001-7609-1939">Michele Cosi</a>)
**Repository:** <a href="https://github.com/LyonsLab/StarBLAST/" target ="_blank">https://github.com/LyonsLab/StarBLAST/</a>
**Version:** 1.0.0a

- [x] **Version:** 1.0.0a

- [x] **Version:** 1.0.0a
