### 2021-01-14 #102 @JasonJWilliamsNY
@oneilsh @AstrobioMike this (#102) is the issue for review. You each have a checklist of steps to make reviewing easier and there are PDFs above where you can the short paper and any changes @CosiMichele  makes in response to your comments. What should follow is really a conversation where everyone participates with some suggestions. If you are unclear about anything please let me know and I will try and help. Once your checklists and reviews are complete I can move forward.

I'll ask our bot Whedon to set an initial reminder for two weeks. Of course, you will have to take the time you need, so as things come up just let everyone know if more time is needed. I think everyone is flexible, realizes the circumstances of our day, and appreciates your contribution as a volunteer. Thanks!
