### 2021-01-26 #102 @AstrobioMike
@JasonJWilliamsNY I'm embarrassed to say I missed the 7-day window to accept the collaboration invitiation, and now I can't interact with my checkboxes :( Any chance we can re-send that?
