### 2021-01-26 #102 @JasonJWilliamsNY
> @JasonJWilliamsNY, just noting I have worked closely with Sateesh Peri (one of the co-authors) doing a few workshops within the past 4 years, though we have not published anything together and do not work at the same organization. I do consider us friends, however, I can assure that this wouldn't affect my ability to provide an impartial assessment of this work, so I request that this conflict be waived. Please let me know if this is ok. Sorry I missed his name in the author list earlier.

Thanks for disclosing @AstrobioMike. As long as you feel you can provide objective review, your disclosure does not appear to meet the COI definition referenced at the top of this issue. You should be fine to proceed but if you have further concerns just let me know.
