### 2021-01-27 ed39cc0
Update 2_StarBLAST-VICE.rst
- just to be clear a blast db doesn't need to be made of an organism
