### 2021-01-27 baf0c68
Update 3_StarBLAST-Docker.rst
- re: step 5 - image might not show up unless "Show All" tab is selected
- re: step 9, i assume this means to repeat steps 4-8?
