### 2021-01-27 f359657
Update 4_StarBLAST-HPC.rst
adding the `cd` first, otherwise `pwd` won't necessarily tell folks their home location
