### 2021-01-27 24128d8
Update paper.md
- adjusted to avoid saying to match "nucleotide" sequences alone, as the end of the sentence was previously worded
- comma added
- thought "addresses a variety of technical abilities" was confusing
