+++ paper.md
@@ -48,7 +48,7 @@ bibliography: paper.bib
 
 # Summary
 
-Basic Local Alignment Search Tool (BLAST) [@madden2002] is a heuristic nucleotide and protein search tool frequently used in biological research to match nucleotide sequences to a genetic database. It is becoming increasingly common in classrooms, including the high school-grade level [@form2011, @shaffer2010] through the  publicly accessible version, NCBI BLAST [@johnson2008]. However, NCBI BLAST can become problematic in a classroom environment, as many job submissions from the same IP address may be throttled (IP block), response time may be slow, and users are unable to import custom databases. In addition, updates to NCBI databases may produce results incongruent with lesson plans. SequenceServer [@priyam2019] addresses some of these needs by providing a student-friendly, web-accessible interface of NCBI-BLAST+ [@camacho2009] and an expanded collection of BLAST tools, with options to use custom databases. However, SequenceServer does not address the scalability issues for large class sizes and deploying the software is still difficult for instructors with minimal computational resources and technical expertise. To address these needs, we developed StarBLAST, a modular deployment strategy for SequenceServer that addresses a variety of technical abilities and classroom sizes. This is achieved by leveraging a master-worker framework for distributed and scalable computing using WorkQueue and Makeflow [@albrecht2012]. WorkQueue manages jobs (compute tasks) to be distributed from a “Master” node to designated “Worker” nodes for completion. Its inherent scalability makes it functional for large and small classrooms. Through CyVerse VICE [@devisetty2016], an instructor can quickly deploy SequenceServer and databases on a web-based interface. Here, we describe the three deployable methods of StarBLAST that account for custom databases, classroom size, technical expertise, and computational resources.
+Basic Local Alignment Search Tool (BLAST) [@madden2002] is a heuristic nucleotide and protein search tool frequently used in biological research to match sequences to those in a database. It is becoming increasingly common in classrooms, including the high school-grade level [@form2011, @shaffer2010] through the  publicly accessible version, NCBI BLAST [@johnson2008]. However, NCBI BLAST can become problematic in a classroom environment, as many job submissions from the same IP address may be throttled (IP block), response time may be slow, and users are unable to import custom databases. In addition, updates to NCBI databases may produce results incongruent with lesson plans. SequenceServer [@priyam2019] addresses some of these needs by providing a student-friendly, web-accessible interface of NCBI-BLAST+ [@camacho2009] and an expanded collection of BLAST tools, with options to use custom databases. However, SequenceServer does not address the scalability issues for large class sizes, and deploying the software is still difficult for instructors with minimal computational resources and technical expertise. To address these needs, we developed StarBLAST, a modular deployment strategy for SequenceServer that is suitable for a variety of technical skill-levels and classroom sizes. This is achieved by leveraging a master-worker framework for distributed and scalable computing using WorkQueue and Makeflow [@albrecht2012]. WorkQueue manages jobs (compute tasks) to be distributed from a “Master” node to designated “Worker” nodes for completion. Its inherent scalability makes it functional for large and small classrooms. Through CyVerse VICE [@devisetty2016], an instructor can quickly deploy SequenceServer and databases on a web-based interface. Here, we describe the three deployable methods of StarBLAST that account for custom databases, classroom size, technical expertise, and computational resources.
 
 # Statement of Need
 
@@ -64,7 +64,7 @@ StarBLAST-VICE is a dockerized image of SequenceServer integrated as a VICE appl
 
 ## StarBLAST-Docker
 
[…]
