### 2021-01-27 #102 @AstrobioMike
Nice work, @CosiMichele et al! Overall the purpose, infrastructure, and documentation seem great :) Thank you for your work on this!

**First I'll comment on specific boxes from the JOSE review checklist that I have yet to check:**

- [ ] **Version: Does the release version given match the GitHub release (v1.0.0)?**
  * No, there are no releases on github, but maybe this is because you are waiting to release once the review process is done?

- [ ] **Example usage: Do the authors include examples of how to use the software?**
  * There is documentation for installation, and usage such as adding a custom database, but I didn't see any example usage documentation for actually using the program – though this is/would be minimal as it's a very clean interface, and I understand the assumption is people know how to use the web blast already. But maybe a short page for general usage could be added to the end of the documentation? Something with images/descriptions of the input and out interfaces?

- [ ] **Tests: Are there automated tests or manual steps described.**
  * Just noting, these don't exist, but i checked this box because I don't think this really applies here.

- [ ] **Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support**
  * No explicit guidelines are currently presented for third parties wishing to contribute, report issues or problems, or seek support. Maybe a page could be added to the documentation providing contact info and/or a link to the github issues page stating these things can be done there?

**I made some minor suggestions/clarifications in this pull request: https://github.com/LyonsLab/StarBLAST/pull/29**
As noted there, I don't consider any of it as required from my point of view.

**Other minor snags/suggestions:**

- For adding our own database, following instructions here: https://starblast.readthedocs.io/en/latest/2_StarBLAST-VICE.html#adding-your-own-databases-to-starblast-vice
    - 3 different ones come up currently when searching for "Create BLAST Database"
       - the first one on the list currently, put up by Alice Minotto, seems to not be the one we want (for this one we need to adjust molecule type in parameters to try to run it, and it ultimately fails)
       - so should specify which should be used better
    - In the instructions to use this database in the same area, it says to select the folder containing our database, but the folder i created for this seems to only hold a directory with the name and date of the "analysis" launched to create the db, and within that a logs directory. I can't seem to see the actual db files. When choosing the directory I created, it still works, but it should be made clear in the docs here that the user shouldn't expect to see the actual files (if that is the norm)

**General notes for the review:**
- I was able to only fully test the StarBLAST-VICE version, due to not having an HPC to utilize for this and not having a current XSEDE allocation open. Maybe @oneilsh will be able to test those, but given my reviewing of the full documentation, my experience with the VICE version, and that those others are getting a little into more technical grounds anyway (particularly the HPC avenue), I don't think it's a big deal if we can't directly independently test them.
