### 2021-02-21 #102 @labarba
Hi @CosiMichele 👋  — Do give us a quick update about your plan for addressing the comments and suggestions contributed by @AstrobioMike.

I've also sent a ping via email to our second reviewer, to learn about his status. These are tough times for most people, and we do our best.
