### 2021-02-22 #102 @oneilsh
Hmm, I was able to accept the new invite, but I'm not able to interact with the checkboxes or see any options for editing the issue comment.
