### 2021-02-22 #102 @labarba
@openjournals/dev --- I have reached a dead-end with a reviewer permissions issue. Could you take a look? The reviewer is unable to edit the checklist, despite being re-invited and accepting the invitation.
