+++ docs/5_SS_usage.rst
@@ -2,17 +2,21 @@
 Using SequenceServer
 ********************
 
-SequenceServer allows to access BLAST+ commands through a simple GUI. Here, we show examples of how to BLAST using SequenceServer. For additional documentation please visit SequenceServer's `official website <https://sequenceserver.com/>` and `original publicaiton <https://academic.oup.com/mbe/article/36/12/2922/5549819>`.
+SequenceServer allows to access BLAST+ commands through a simple GUI. Here, we show examples of how to BLAST using SequenceServer. For additional documentation please visit SequenceServer's `official website <https://sequenceserver.com/>`_ and `original publicaiton <https://academic.oup.com/mbe/article/36/12/2922/5549819>`_.
 
 .. note::
 
    These examples will take into consideration that you already have launched StarBLAST. Visit the other User Guides to know more on launching StarBLAST.
 
 
-Sequenceserver Main Page
-========================
+SequenceServer's Main Page
+==========================
 
 On the main page, the user will see the main input box where nucleotide (DNA) or amino acid (protein) sequences can be input using the FASTA convention. 
 Below the input box, there are the nucleotide databases (left) and protein databases (right). 
 The user will be able to choose which databases to BLAST against by clicking the boxes left to the databases' names. 
-Under the databases, the user can specify advanced parameters before BLASTing. A list and description of all the advanced options can be accessed by pressing the "?" button. These advanced parameters can heavily influence the resulting BLAST results, we suggest to read the descriptions beforehand.
+Under the databases, the user can specify advanced parameters before BLASTing. A list and description of all the advanced options can be accessed by pressing the "**?**" button. 
+
+.. note::
+  
+  Advanced Parameters can heavily influence the resulting BLAST results, we suggest to read the descriptions beforehand.
