### 2021-02-22 #102 @oneilsh
Hi @CosiMichele, @JasonJWilliamsNY,

I'll add some comments here while I work on getting checkbox access :)

First, I think this is an excellent effort - bioinformatics teaching is challenging, particularly given the heavy resource needs for most bioinformatics software, the very bursty nature of teaching those methods, and how difficult such software is to manage. I appreciate the multiple deployment options presented, though I think the DE VICE version is the most exciting, providing an easy on-ramp for those without HPC resources at their disposal. I also appreciate that this project doesn't re-invent wheels, utilizing SequenceServer which is a great piece of software.

Some comments on boxes that I'm not sure I can check right away:

**Release version: **
As mentioned by the other reviewer, there are no 'releases' listed on GitHub, but I'm not sure that makes sense for the VICE deployment (which I think is managed inside the DE?) I confess I'm not too familiar with GitHub -> Docker release machinery, but you might find https://docs.github.com/en/actions/guides/publishing-docker-images useful for that aspect. Since a "release" of this package might involve updates to more than one docker image (master and worker), a strategy for tagging (with semantic versioning?) those images and tying them to the repo-described version (v1.0.0 here) would be wise I think.

**Installation: **
I also don't have access to JetStream or an HPC infrastructure with the required dependencies, so I was only able to test the DE VICE version. It worked - I think the screenshots in the JetStream deployment docs help a lot, having something comparable (or a walk-through video) would help with the DE version since CyVerse DE isn't straightforward for those not used to it. I had trouble opening the app for example until I figured out which DE UI element did the trick.

The HPC documentation is technical and terse, but on first blush seems complete enough, assuming a compatible environment is available. That I think is hard to achieve though: I have access to multiple "HPC" systems, but none with the specific set of pre-reqs suggested (modules properly configured, iRODS, cctools, pbs or similar batch-queue, and user-level ability to run docker). If targeting batch clusters is a strong goal, I would suggest looking into what few common stacks are available - OpenOnDemand is one possibility, but beyond a 1.0 release I suspect. The docs in the HPC section suggest iRODS as a requirement, but I see there are non-iRODS docker images available but their usage is glossed over.

**Example Usage: **
I think examples of usage for instructors in particular would be most helpful. Many would come from using NCBI's GUI and have no experience setting up infrastructure. The SequenceServer docs fortunately provide a good reference for newcomers and could be selectively referenced. References for signing up for CyVerse would also be useful. It looks like with the DE version I can link to results via URL - how long do those links last? Could students safely submit their work by providing those links in e.g. homework submissions? Suggested workflows for incorporating StarBLAST into the classroom would go a long way to reaching the audience I think, especially if they include installation as well as usage walk-throughs with screenshots or video.

**Community Guidelines: **
Contribution guidance isn't given, but I think this could be addressed easily by suggesting using Issues for issues or questions and possibly a "seeking help" section of the README with suggestions (e.g. training materials leveraging the tools or guides for customization). Some extra documentation for navigating the codebase might be appreciated by those wanting to dive in to customizations or make improvements.
