+++ docs/2_StarBLAST-VICE.rst
@@ -12,59 +12,55 @@ StarBLAST-VICE is launchable with a maximum of 8 CPU cores, 16 GB RAM, and 512 G
 Launching StarBLAST-VICE with Example Databases
 ===============================================
 
-**1.** Click on the following button to launch SequenceServer in CyVerse Discovery Environment with two BLAST databases (Human_GRCh38_p12 & Mouse_GRCm38_p4 - **requires CyVerse account**).
+(1) Click on the following button to launch SequenceServer in CyVerse Discovery Environment with two BLAST databases (Human_GRCh38_p12 & Mouse_GRCm38_p4 - **requires CyVerse account**).
 
 	|seqserver_QL|_
 	
 
-**2.** Choose your own analysis name and the DE output folder. Click "Launch Analysis".
+(2) Choose your own analysis name and the DE output folder. Click "Launch Analysis".
 
 
-**3.** Check the notifications Bell Icon for a link to access your SequenceServer instance. This might take a few minutes.
+(3) Check the notifications Bell Icon for a link to access your SequenceServer instance. This might take a few minutes.
 
 
-**4.** To test, click `here <https://www.ncbi.nlm.nih.gov/nuccore/NG_007114.1?from=4986&to=6416&report=fasta>`_ for a sample DNA sequence.
+(4) To test, click `here <https://www.ncbi.nlm.nih.gov/nuccore/NG_007114.1?from=4986&to=6416&report=fasta>`_ for a sample DNA sequence.
 
 
-**5.** Paste the query sequence, select both the available databases and submit job.
+(5) Paste the query sequence, select both the available databases and submit job.
 
 Adding Your Own Databases to StarBLAST-VICE
 ===========================================
 
 To add your own BLAST databases you will need a :code:`.fasta (or .fa, .faa, .fna)`  file containing the reference sequences you'd like to use. These are easily aquirable from NCBI or other databases.
 
-**1.** Within the CyVerse DE, click on the "Data" icon. 
+(1) Within the CyVerse DE, click on the "Data" icon. 
 
+(2) Select "Upload" and specify the import from your Desktop or its URL. This will be stored in your personal folder.
 
-**2.** Select "Upload" and specify the import from your desktop or its URL. This will be stored in your personal folder.
+(3) Click on the "Apps" icon and use the search bar to find "Create BLAST Database" or click `here <https://de.cyverse.org/de/?type=apps&app-id=decdd668-5616-11e7-9724-008cfa5ae621&system-id=de>`_. As there is a possiblity that there may be multiple apps with the same name, please locate the "Create BLAST Database" developed by developed by Upendra Kumar Devisetty (use the better reviewed one).
 
+(4) Enter a name for your database under "Analysis Name"; this will become your database containing folder. 
 
-**3.** Select "File" + "New Folder" to create a folder for your database.
+(4.1) "Select output folder" should be your personal folder or any folder of your choice (default will be a foler named "analyses" within your personal folder).
 
+(4.2) In the "Inputs" tab, select "Browse" and choose the fasta file you uploaded. Select Nucleotide or Protein under "Input Sequence Format". Under "Prefix" choose a name to well reflect your database (e.g. :code:`a_thaliana`).
 
-**4.** Click on the "Apps" icon and use the search bar to find "Create BLAST Database" or click `here <https://de.cyverse.org/de/?type=apps&app-id=decdd668-5616-11e7-9724-008cfa5ae621&system-id=de>`_. As there is a possiblity that there may be multiple apps with the same name, please locate the "Create BLAST Database" developed by developed by Upendra Kumar Devisetty.
-
-
[…]
