### 2021-02-23 #102 @CosiMichele
Hi everyone, @AstrobioMike , I believe I have addressed all the issues which you have brought us:

1. Although we believe that this is StarBLAST 1.0, we will be awaiting for this review to pass before assigning the version number. However, if required, we will be happy to add the version to our GitHub.
2. We have added a page for usage and examples (https://starblast.readthedocs.io/en/latest/5_SS_usage.html), including images from our StarBLAST-Docker release. Please let us know if this is satisfying.
3. A small section (Contacts & Issues) has been added at the bottom of our rtd homepage, linking our GitHub issues page.
4. I understand that the process to add a Database to StarBLAST-VICE was unclear; I've modified that section to better reflect steps to undertake in order to add a personalized database. Again, if more needs to be added, let us know.

Additionally, I have merged the small edits suggested in the pull request.

Thank you @AstrobioMike for the feedback!
Finally you @labarba and @oneilsh for checking on this. Let us know what else we can do!
