### 2021-03-01 #102 @arfon
> @openjournals/dev --- I have reached a dead-end with a reviewer permissions issue. Could you take a look? The reviewer is unable to edit the checklist, despite being re-invited and accepting the invitation.

Could you tell me what browser you're using?
