### 2021-03-03 #102 @CosiMichele
Hi everyone, thank you once again for your time and patience.

Thank you @AstrobioMike for taking time and reviewing StarBLAST! Your input has really helped us.

Thank you @oneilsh for your comments! Here are some updates:

1. Since both you and @AstrobioMike pointed out the version release, I went ahead and added it to the GitHub code.
2. I have edited the HPC requirements installation, addressing how to install iRODS (and ncbi-BLAST+, CCTools) for non users. Hopefully this will be helpful.
3. I had added example usage and community guidelines in my previous edit and highlighted in my previous comment - let us know if further changes are required.

Once again, thank you all for your time and feedback. We look forward to hearing from you!
