### 2021-03-05 #102 @oneilsh
@arfon - Chrome, 88.0.4324.192 (Official Build) (x86_64)

@CosiMichele nice updates! They do address most of my thoughts. My only remaining suggestion would be to include some screenshots on the installation/access of the DE version.
