+++ docs/2_StarBLAST-VICE.rst
@@ -12,19 +12,26 @@ StarBLAST-VICE is launchable with a maximum of 8 CPU cores, 16 GB RAM, and 512 G
 Launching StarBLAST-VICE with Example Databases
 ===============================================
 
-(1) Click on the following button to launch SequenceServer in CyVerse Discovery Environment with two BLAST databases (Human_GRCh38_p12 & Mouse_GRCm38_p4 - **requires CyVerse account**).
+(1) Click on the following button to launch SequenceServer in CyVerse Discovery Environment with the SWISS-PROT protein database (**requires CyVerse account**). If you are already in the DE, you can navigate to the StarBLAST-VICE app through the Apps button and searching for "StarBLAST-VICE".
 
-	|seqserver_QL|_
+|seqserver_QL|_
 	
+.. note::
+
+   SWISS-PROT is a curated protein sequence database, read more on the `release <https://www.uniprot.org/statistics/Swiss-Prot>`_ or its `orignal publication <https://www.ncbi.nlm.nih.gov/pmc/articles/PMC102476/>`_ 
+
+|DE_running_tut_01|
 
 (2) Choose your own analysis name and the DE output folder. Click "Launch Analysis".
 
+|DE_running_tut_02|
 
-(3) Check the notifications Bell Icon for a link to access your SequenceServer instance. This might take a few minutes.
+(3) Check the notifications Bell Icon for a link to access your SequenceServer instance. This might take a few minutes. Once the notification shows that the app is running, click on the link. This will open a loading screen in a new tab; Once the app is loaded, you should be able to BLAST through the SequenceServer app.
 
+|DE_running_tut_03|
+|DE_running_tut_04|
 
 (4) To test, click `here <https://www.ncbi.nlm.nih.gov/nuccore/NG_007114.1?from=4986&to=6416&report=fasta>`_ for a sample DNA sequence.
-
 
 (5) Paste the query sequence, select both the available databases and submit job.
 
@@ -37,13 +44,21 @@ To add your own BLAST databases you will need a :code:`.fasta (or .fa, .faa, .fn
 
 (2) Select "Upload" and specify the import from your Desktop or its URL. This will be stored in your personal folder.
 
+|DE_running_tut_05|
+
 (3) Click on the "Apps" icon and use the search bar to find "Create BLAST Database" or click `here <https://de.cyverse.org/de/?type=apps&app-id=decdd668-5616-11e7-9724-008cfa5ae621&system-id=de>`_. As there is a possiblity that there may be multiple apps with the same name, please locate the "Create BLAST Database" developed by developed by Upendra Kumar Devisetty (use the better reviewed one).
+
+|DE_running_tut_06|
 
 (4) Enter a name for your database under "Analysis Name"; this will become your database containing folder. 
 
    1. "Select output folder" should be your personal folder or any folder of your choice (default will be a foler named "analyses" within your personal folder).
 
[…]
