### 2021-03-08 c502753
VICE tutorial & cleanup
