### 2021-03-08 #102 @CosiMichele
@oneilsh Thank you for the suggestion! I have added screenshots to help users navigate the DE.

Please let us know if there's anything else we should do/change.
