### 2021-03-30 #102 @CosiMichele
Hi @JasonJWilliamsNY, thanks for checking in!

I have made the changes proposed by @AstrobioMike and @oneilsh to the documentation. I believe there's only one last proposition from @oneilsh, which has been addressed. As things stand, I am just waiting for feedback on these last changes. Thanks again for checking in and all the reviewers' work!
