### 2021-04-14 #102 @JasonJWilliamsNY
@CosiMichele I have reviewed the proof and things look good. The next step is for you to create an archive on either https://zenodo.org/ or https://figshare.com/ and obtain a DOI (which you'll need to give to us). I can then move to the final acceptance and publication steps.

Tip: Here is a nice guide on doing the above with Zenodo: https://guides.github.com/activities/citable-code/#:~:text=Making%20Your%20Code%20Citable%201%20Choose%20your%20repository.,a%20shiny%20new%20badge%20showing%20your%20new%20DOI%21
