### 2021-04-14 #102 @labarba
Hi @CosiMichele 👋
I see we're getting to the end with this one 🎉

We like to ask that the Zenodo metadata be edited to match the JOSE paper: that is match the title, and list of authors (note that the auto-sync with Zenodo lists anyone who has made a commit as author.) Could you make that change?
