### 2021-04-14 #102 @labarba
Should the version be marked as `1.0.0a`?
