### 2021-04-14 #102 @CosiMichele
@labarba , thanks for checking in! I believe I changed the Zenodo metadata to match the JOSE metadata (including the version). Can you please check and let us know if there's more to chance? Thank you!
