### 2021-04-27 #102 @labarba
Congratulations, @CosiMichele, your JOSE paper is published 🚀

Huge thanks to the Editor: @JasonJWilliamsNY and Reviewers: @oneilsh, @AstrobioMike for their contribution—this experiment in alternative publishing is possible thanks to you!
